Tweak: Preview result in AI CSS feature [ED-14993]
