+++ modules/ai/assets/js/editor/page-content.js
@@ -16,7 +16,7 @@ import useUpgradeMessage from './hooks/use-upgrade-message';
 import UsageMessages from './components/usage-messages';
 import { Box, Typography } from '@elementor/ui';
 import Loader from './components/loader';
-import { useEffect, useState } from 'react';
+import { useEffect, useRef, useState } from 'react';
 import { useRequestIds } from './context/requests-ids';
 import { FREE_TRIAL_FEATURES_NAMES } from './helpers/features-enum';
 
@@ -52,6 +52,9 @@ const PageContent = (
 	}, [ isLoading, initialUsagePercentage, isInitUsageDone, updateUsagePercentage ] );
 
 	const { showBadge } = useUpgradeMessage( { usagePercentage, hasSubscription } );
+	const [ sxStyle, setSxStyle ] = useState( { pointerEvents: 'none' } );
+	const timeoutRef = useRef( null );
+
 	const promptDialogStyleProps = {
 		sx: {
 			zIndex: 170000, // Make sure the dialog is above wp attachment details view
@@ -73,6 +76,29 @@ const PageContent = (
 				m: 0,
 				maxHeight: 'media' === type ? '95vh' : '76vh',
 				height: 'auto',
+			},
+		},
+	};
+
+	const codePromptDialogStyleProps = {
+		sx: {
+			...promptDialogStyleProps.sx,
+			...sxStyle,
+		},
+		PaperProps: {
+			onMouseEnter: () => {
+				clearTimeout( timeoutRef.current );
+				setSxStyle( { pointerEvents: 'all' } );
+			},
+			onMouseLeave: () => {
+				clearTimeout( timeoutRef.current );
+				timeoutRef.current = setTimeout( () => {
+					setSxStyle( { pointerEvents: 'none' } );
+				}, 200 );
+			},
+			sx: {
+				...promptDialogStyleProps.PaperProps.sx,
+				pointerEvents: 'auto',
 			},
 		},
 	};
@@ -163,7 +189,7 @@ const PageContent = (
 
 	if ( 'code' === type ) {
 		return (
-			<PromptDialog onClose={ onClose } { ...promptDialogStyleProps }>
+			<PromptDialog onClose={ onClose } { ...codePromptDialogStyleProps }>
 				<PromptHistoryProvider historyType={ HISTORY_TYPES.CODE }>
 					<PromptHistoryActionProvider>
 						<PromptDialog.Header onClose={ onClose }>
